--- @twc_framework/addons/twc_ui/DefaultVehicleSystemsDisplayManager.hpp
@@ -1,47 +1,0 @@
-class DefaultVehicleSystemsDisplayManagerLeft
-{
-	componentType = "EmptyDisplayComponent"; //mandatory
-	/* forcedDisplay = "EmptyDisplay";
-	x = (safezoneX + 0.5 * (((safezoneW / safezoneH) min 1.2) / 40));
-	y = (safezoneY + safezoneH - 21 * ((((safezoneW / safezoneH) min 1.2) / 1.2) / 25));
-	left = 1;
-	right = 0;
-
-	class Components {
-		class SlingLoadDisplay { componentType = "EmptyDisplayComponent"; };
-		class SensorsDisplay { componentType = "EmptyDisplayComponent"; };
-		class UAVFeedDisplay { componentType = "EmptyDisplayComponent"; };
-		class VehicleDriverDisplay { componentType = "EmptyDisplayComponent"; };
-		class VehicleGunnerDisplay { componentType = "EmptyDisplayComponent"; };
-		class VehicleCommanderDisplay { componentType = "EmptyDisplayComponent"; };
-		class MissileDisplay { componentType = "EmptyDisplayComponent"; };
-		class MinimapDisplay { componentType = "EmptyDisplayComponent"; };
-		class CrewDisplay { componentType = "EmptyDisplayComponent"; };
-		class MineDetectorDisplay { componentType = "MineDetectorDisplayComponent"; };
-		class EmptyDisplay { componentType = "EmptyDisplayComponent"; };
-	}; */
-};
-
-class DefaultVehicleSystemsDisplayManagerRight
-{
-	componentType = "EmptyDisplayComponent";
-	/* forcedDisplay = "EmptyDisplay";
-	y = (safezoneY + safezoneH - 21 * ((((safezoneW / safezoneH) min 1.2) / 1.2) / 25));
-	x = (safezoneX + safezoneW) - ((10 * (((safezoneW / safezoneH) min 1.2) / 40)) + 0.5 * (((safezoneW / safezoneH) min 1.2) / 40));
-	left = 0;
-	right = 1;
-
-	class Components {
-		class SlingLoadDisplay { componentType = "EmptyDisplayComponent"; };
-		class SensorsDisplay { componentType = "EmptyDisplayComponent"; };
-		class UAVFeedDisplay { componentType = "EmptyDisplayComponent"; };
-		class VehicleDriverDisplay { componentType = "EmptyDisplayComponent"; };
-		class VehicleGunnerDisplay { componentType = "EmptyDisplayComponent"; };
-		class VehicleCommanderDisplay { componentType = "EmptyDisplayComponent"; };
-		class MissileDisplay { componentType = "EmptyDisplayComponent"; };
-		class MinimapDisplay { componentType = "EmptyDisplayComponent"; };
-		class CrewDisplay { componentType = "EmptyDisplayComponent"; };
-		class MineDetectorDisplay { componentType = "MineDetectorDisplayComponent"; };
-		class EmptyDisplay { componentType = "EmptyDisplayComponent"; };
-	}; */
-};
