+++ @twc_framework/addons/twc_ui/config.cpp
@@ -19,29 +19,9 @@ class CfgPatches {
 	};
 };
 
-class CfgFunctions {
-	class TWC_UI {
-		tag = "TWC_UI";
-		class init {
-			file = "TWC_UI\functions";
-			class init { postInit = 1; };
-			
-			class addDisplay {};
-			class getVisualBearingLocal {};
-			class hideBearing {};
-			class shouldDisplay {};
-			class removeDisplay {};
-			class showBearing {};
-			class hudReset {};
-			
-			/** adds hide functionality to ace spectator camera **/
-			class handleSpectateCompass {};
-			class toggleSpectateCompass {};
-		};
-	};
-};
-
-#include "RscTitles.hpp"
-#include "DefaultVehicleSystemsDisplayManager.hpp"
+#include "CfgFunctions.hpp"
 #include "RadioProtocolBase.hpp"
-#include "ACE_Spectator_Display.hpp"
+
+#include "ui\ACE_Spectator_Display.hpp"
+#include "ui\DefaultVehicleSystemsDisplayManager.hpp"
+#include "ui\RscTitles.hpp"
