+++ @twc_framework/addons/twc_ui/ui/DefaultVehicleSystemsDisplayManager.hpp
@@ -1,0 +1,7 @@
+class DefaultVehicleSystemsDisplayManagerLeft {
+	componentType = "EmptyDisplayComponent";
+};
+
+class DefaultVehicleSystemsDisplayManagerRight {
+	componentType = "EmptyDisplayComponent";
+};
